Создан MovieQuizPresenter, перенесён метод convert, устранено дублирование данных.
